# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  POS-fiplex_internal_common.ErrorResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
    type: object
  POS-fiplex_internal_common.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  internal_settings.BrandingSettingsResponse:
    properties:
      app_logo:
        type: string
      app_name:
        type: string
      footer_text:
        type: string
    type: object
  internal_settings.PrinterSettingsResponse:
    properties:
      auto_print:
        type: boolean
      connection:
        type: string
      paper_width:
        type: string
      print_method:
        type: string
    type: object
  internal_settings.UpdateBrandingRequest:
    properties:
      app_logo:
        type: string
      app_name:
        maxLength: 100
        minLength: 3
        type: string
      footer_text:
        maxLength: 200
        type: string
    required:
    - app_name
    type: object
  internal_settings.UpdatePrinterSettingsRequest:
    properties:
      auto_print:
        type: boolean
      connection:
        type: string
      paper_width:
        enum:
        - 58mm
        - 80mm
        type: string
      print_method:
        enum:
        - BE
        - FE
        type: string
    required:
    - auto_print
    - connection
    - paper_width
    - print_method
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: POS Fiplex API
  title: POS Fiplex API
  version: "1.0"
paths:
...
  /settings/branding:
    get:
      consumes:
      - application/json
      description: 'Retrieve branding settings (app name, logo, footer text, theme
        colors) for the application (Roles: authenticated)'
      produces:
      - application/json
      responses:
        "200":
          description: Branding settings fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_settings.BrandingSettingsResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: Get branding settings
      tags:
      - Settings
      x-roles:
      - admin
      - manager
      - cashier
    put:
      consumes:
      - application/json
      description: 'Update application branding settings (Roles: admin)'
      parameters:
      - description: Branding update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_settings.UpdateBrandingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Branding settings updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_settings.BrandingSettingsResponse'
              type: object
        "400":
          description: Invalid request body or validation failure
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: Update branding settings
      tags:
      - Settings
      x-roles:
      - admin
  /settings/branding/logo:
    post:
      consumes:
      - multipart/form-data
      description: 'Upload and update the application logo image (Roles: admin)'
      parameters:
      - description: Logo image file
        in: formData
        name: logo
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Logo updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Logo file is required or file size exceeds limit
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: Update app logo
      tags:
      - Settings
      x-roles:
      - admin
  /settings/printer:
    get:
      consumes:
      - application/json
      description: 'Retrieve printer settings like connection string and paper width
        (Roles: authenticated)'
      produces:
      - application/json
      responses:
        "200":
          description: Printer settings fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_settings.PrinterSettingsResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: Get printer settings
      tags:
      - Settings
      x-roles:
      - admin
      - manager
      - cashier
    put:
      consumes:
      - application/json
      description: 'Update global printer configuration (Roles: admin)'
      parameters:
      - description: Printer settings update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_settings.UpdatePrinterSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Printer settings updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_settings.PrinterSettingsResponse'
              type: object
        "400":
          description: Invalid request body or validation failure
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: Update printer settings
      tags:
      - Settings
      x-roles:
      - admin
...
swagger: "2.0"